requirements.txt
Flask==2.3.2
pymongo==4.15.1
cryptography==41.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
